+++ src/reports/events/handlers/report-deleted.handler.ts
@@ -1,6 +1,6 @@
 import { EventsHandler, IEventHandler } from "@nestjs/cqrs";
 import { ReportDeletedEvent } from "../impl/report-deleted.event";
-import { Logger } from "@nestjs/common";
+import { Logger, NotFoundException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { ReportDto } from "reports/dtos/reports.dto";
 import { Repository } from "typeorm";
@@ -10,13 +10,19 @@ export class ReportDeletedHandler implements IEventHandler<ReportDeletedEvent> {
   constructor(
     @InjectRepository(ReportDto)
     private readonly repository: Repository<ReportDto>
-  ) {}
+  ) { }
 
   async handle(event: ReportDeletedEvent) {
+    Logger.log(event.reportId, "ReportDeletedEvent");
+    const { streamId, reportId } = event;
+
     try {
-      Logger.log(event.reportId, "ReportDeletedEvent");
-      const reportId = event.reportId;
-      return await this.repository.delete({ _id: reportId });
+      const report = await this.repository.findOne({ _id: reportId });
+      if (report) {
+        await this.repository.delete({ _id: reportId });
+        return;
+      }
+      throw new NotFoundException(`Report with _id ${reportId} does not exist.`);
     } catch (error) {
       Logger.error(error, "", "ReportDeletedEvent");
     }
